Allow blogger to write articles in wysiwyg
The WYSIWYG editor ckeditor has been implemented and bloggers now have
the option to write blogs in either markdown or wysiwyg.

This has been applied in the creation of blog posts and the display of
them there after. Some additions had to be made to the article database
to accommodate this change.