Refactor home page for current design
The display of articles has been altered and the whole of the blogger
feature spec has been changed to reflect the new designs.

Added sign in/up pages and sign in to nave bar for use on certain page
widths, this will later be controlled by the css media queries.