Reset CSS ready for ITCSS implementation
The current CSS has been removed temporarily so that it can be
implemented in the ITCSS convention.

Folders have been added for the different specificity stages that will
be used and a normalize.css has been added to reset browser applied
stylings. The fonts.scss has also been created and comments containing
the different syntax for the fonts used has been added for ease of
accessibility.