Add styling to the single article page
Stylings have been added to the whole of the article page. The article,
comments and the links to other user articles has been completed.

The Javascript has also been removed for editing articles and adding
comments. This has been removed and replaced with CSS checkbox
transitions (if that is the right term for them).

Obviously with using existing classes they may have been changed
slightly to reflect these changes.

Another method has been added to the comment class to return the time
since the comment was posted in second, minutes, ... , years. This
method is then used in the comments table to display nicely.